Safari-611.1.10/LayoutTests/fast/dom/Geolocation/dataURL-watchPosition-expected.txt

PASS error.code is error.PERMISSION_DENIED.
PASS error.message is "Origin does not have permission to use Geolocation service".
